backend/requirements.txt
# Core web framework
flask==2.3.3
flask-cors==4.0.0
flask-restful==0.3.10

# NLP and text processing
spacy==3.7.2
nltk==3.8.1
pandas==2.1.3
numpy==1.24.3

# PDF and document processing
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11

# File handling and utilities
python-magic==0.4.27
werkzeug==2.3.7

# Data validation
pydantic==2.5.0

# HTTP requests (for job board APIs)
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Machine learning (for job matching)
scikit-learn==1.3.2
scipy==1.11.4

# Logging and monitoring
structlog==23.2.0

# Database (for future use)
sqlalchemy==2.0.23